# mobile/swagger.json
{
    "schemes": [],
    "swagger": "2.0",
    "info": {
        "description": "",
        "title": "",
        "contact": {},
        "version": ""
    },
    "host": "",
    "basePath": "",
    "paths": {
        "/inventory": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取库存列表，支持按产品变体和店铺筛选，并支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "库存管理"
                ],
                "summary": "获取库存列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "产品变体ID",
                        "name": "product_variant_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "店铺ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "每页数量，默认10",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取库存列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Inventory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的库存记录，需提供产品变体ID、店铺ID和库存数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "库存管理"
                ],
                "summary": "创建库存",
                "parameters": [
                    {
                        "description": "库存信息",
                        "name": "inventory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Inventory"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/models.Inventory"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "产品变体或店铺不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "库存记录已存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/inventory/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取库存的详细信息，包括产品变体和店铺信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "库存管理"
                ],
                "summary": "获取库存详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "库存ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取库存详情",
                        "schema": {
                            "$ref": "#/definitions/models.Inventory"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "库存不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新现有库存记录，可以修改库存数量并提供更新原因",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "库存管理"
                ],
                "summary": "更新库存",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "库存ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "库存信息",
                        "name": "inventory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/models.Inventory"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "库存不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定ID的库存记录，此操作不可逆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "库存管理"
                ],
                "summary": "删除库存",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "库存ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "库存不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据"
                },
                "error": {
                    "description": "错误信息，仅在发生错误时返回",
                    "type": "string",
                    "example": "错误信息"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "操作成功"
                }
            }
        },
...
        "models.Dictionary": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "字典类型编码",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "description": "关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DictionaryItem"
                    }
                },
                "name": {
                    "description": "字典类型名称",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：启用/禁用",
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.DictionaryItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "字典项编码",
                    "type": "string"
                },
                "color": {
                    "description": "颜色值（用于前端展示）",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "dictionary": {
                    "description": "关联",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Dictionary"
                        }
                    ]
                },
                "dictionaryCode": {
                    "description": "关联的字典类型编码",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "字典项名称",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：启用/禁用",
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "description": "字典项值",
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "description": "详细错误信息，仅在开发环境返回",
                    "type": "string"
                },
                "message": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
        "models.Inventory": {
            "type": "object",
            "properties": {
                "alertQuantity": {
                    "description": "预警数量",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastCheckTime": {
                    "description": "最后盘点时间",
                    "type": "string"
                },
                "productVariant": {
                    "description": "商品变体",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ProductVariant"
                        }
                    ]
                },
                "productVariantID": {
                    "description": "商品变体ID",
                    "type": "integer"
                },
                "quantity": {
                    "description": "库存数量",
                    "type": "integer"
                },
                "store": {
                    "description": "关联",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Store"
                        }
                    ]
                },
                "storeID": {
                    "description": "店铺ID",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "数据项列表"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "每页记录数",
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "models.ProductVariant": {
            "type": "object",
            "properties": {
                "Barcode": {
                    "description": "条形码",
                    "type": "string"
                },
                "Color": {
                    "description": "颜色",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.DictionaryItem"
                        }
                    ]
                },
                "ColorID": {
                    "description": "颜色ID",
                    "type": "integer"
                },
                "CostPrice": {
                    "description": "成本价（可能与主商品不同）",
                    "type": "number"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "Fabric": {
                    "description": "面料",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.DictionaryItem"
                        }
                    ]
                },
                "FabricID": {
                    "description": "面料ID",
                    "type": "integer"
                },
                "ID": {
                    "type": "integer"
                },
                "ProductID": {
                    "description": "关联的商品ID",
                    "type": "integer"
                },
                "QRCode": {
                    "description": "二维码数据",
                    "type": "string"
                },
                "RetailPrice": {
                    "description": "零售价（可能与主商品不同）",
                    "type": "number"
                },
                "SKU": {
                    "description": "变体SKU",
                    "type": "string"
                },
                "Season": {
                    "description": "季节",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.DictionaryItem"
                        }
                    ]
                },
                "SeasonID": {
                    "description": "季节ID",
                    "type": "integer"
                },
                "Size": {
                    "description": "尺码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.DictionaryItem"
                        }
                    ]
                },
                "SizeID": {
                    "description": "尺码ID",
                    "type": "integer"
                },
                "Status": {
                    "description": "状态：启用/禁用",
                    "type": "boolean"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Store": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "manager": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "type": {
                    "description": "男装店/女装店",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}